Align FFI bindings with ffigen output; upgrade boringssl_dart
`posix_linux.g.dart` and `windows_kernel32.g.dart` are now exactly
what ffigen produces on Linux / Windows runners (glibc internal
parameter names, native C types like UnsignedInt / WChar / UnsignedLong,
the `_SECURITY_ATTRIBUTES` opaque struct, decimal-form constants).
`hardlink.dart`'s `cast<Uint16>()` becomes `cast()` so the inferred
`Pointer<WChar>` matches the new binding signatures. `pubspec.lock`
bumps boringssl_dart to pick up the Windows build-hook fix.

+++ lib/src/ffi/hardlink.dart
@@ -97,11 +97,7 @@ void _windowsLink(String source, String target) {
   final newName = target.toNativeUtf16();
   final existing = source.toNativeUtf16();
   try {
-    final ok = win32.CreateHardLinkW(
-      newName.cast<Uint16>(),
-      existing.cast<Uint16>(),
-      nullptr,
-    );
+    final ok = win32.CreateHardLinkW(newName.cast(), existing.cast(), nullptr);
     if (ok == 0) {
       final err = win32.GetLastError();
       throw IoError(
